Move config to external
